Updated S2D code with version last modified on 11/28/2018.  It includes the following changes:
1. Modification of the fugacity function to include constant background ambient air concentration in ug/m3

2. Changes to other functions so that all houses run, even ones that use no products containing the chemical

This version of the code uses a modified chemical properties file that has columns for GM and GSD that define the distribution of background air concentration (chemical_properties_zero.csv). Currently, we have set these values to 0 for all chemicals.  If there is a non-zero background air concentration, then GM must be positive and GSD must exceed 1.