Add files via upload
This file is an updated chemical properties file associated with the S2D code last modified on 11/28/2018 that includes background ambient air concentration.  The file has columns for GM and GSD that define the distribution of background air concentration.  Currently, we have set these values to 0 for all chemicals.  If there is a non-zero background air concentration, then GM must be positive and GSD must exceed 1.